moved the auth route logic to a yield container/component
this allows to keep the navbar and footer independent from the yield
-Navbar maybe need to be included.....
