+++ client/modules/core/routes.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {mount} from 'react-mounter';
 
-import MainLayout from './containers/main_layout.js';
+import MainLayout from './components/main_layout.jsx';
 import Home from './components/home.jsx';
 import Login from './components/login.jsx';
 import Profile from './components/profile.jsx';
